+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_ca.dart
@@ -664,4 +664,7 @@ class StreamChatLocalizationsCa extends GlobalStreamChatLocalizations {
 
   @override
   String pollSomeoneCreatedText(String username) => '$username ha creat';
+
+  @override
+  String get draftLabel => 'Esborrany';
 }