+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_es.dart
@@ -666,4 +666,7 @@ No es posible añadir más de $limit archivos adjuntos
 
   @override
   String pollSomeoneCreatedText(String username) => '$username ha creado';
+
+  @override
+  String get draftLabel => 'Borrador';
 }