+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_ko.dart
@@ -640,4 +640,7 @@ class StreamChatLocalizationsKo extends GlobalStreamChatLocalizations {
 
   @override
   String pollSomeoneCreatedText(String username) => '$username님이 생성했습니다';
+
+  @override
+  String get draftLabel => '임시글';
 }